pytorch/requirements.txt
dlib
visdom
numpy>=1.17.0
Pillow>=6.1.0
scipy>=1.3.0
six>=1.12.0
torch>=1.1.0
torchvision>=0.3.0
jsonpatch
torchtoolbox
matplotlib
imutils
scikit-image